Fix docker can't properly see config.properties changes from $CONFIG_PATH
We need to bind a directory with the config.properties file instead of the file itself to allow docker to properly see changes on the file.

Previously, depending on the way the file was edited, docker could still see the old file instead of the one with the new changes.