Drop USER_ID and GROUP_ID as they cause problems with the host
It seems difficult to link appuser inside the container with an user in the host. The idea with this was to provide a way to protect the config path on the host, but I just need to try another solution. In any case, this should not be something we need to handle inside the project, so I am removing them to simplify.